Overhaul Work In Progress
Most of the activities and classes that were planned for Quiz version of
the app were re-written and re-designed to match the new concept of the
application.
New Database table was created for the quotes and new php file that will
serve as a connection between the app and the server to insert all the
data into the Database.
User class was created as a Singleton and all required elements for the
user were put inside.
There could be additional class for the Quote but after careful
consideration I decided that at this stage of the application it is not
necessary and this might be included in user.
Adding quotes is working fine. User class creates a separate list for
each logged in user and this list is loaded and saved for every user
separately. Purely for the purpose of each user having its own personal
list. Proper solution for this issue would be to create a direct
connection with the Database. This will be most likely implemented in
future.
Currently working on the editable list of quotes and timer for changing
"quote of the day"