Main Part of Genereal Overhaul Completed
Most of the functionality and the desgin of the new version of teh
application is  already in place.
Refactored the whole appl;ication with looking at the comments, methods,
clases and variables.
Removed unnecessary functionality.

Implemented new version of the Main Home page where user see changing
quotes and can set the change time for them. Added new activity where
user can add quotes which are sent to the database. Added another
activity that displays all the quotes for the current user.
Addedd constant notification that displays the quote and time left to
change.
Started to work on Foreground Service instead of simple notification.