Final Submission For The SET08114 Coursework
As in the title this is the final submission for the coursework.
Content of the app was corrected whenever any mistakes, errors or issues
were found. The same applies to the layouts.
All classes were fully commented and descriptions were filled in.
All methods were checked for correctness and basic testing was
performed. Found bugs were removed.
Service class was abbandonned and new Notification class was created
instead. Added new functionality of clicking on notification that
restores the main Home activity.